server/requirements.txt
litestar[standard]
tinydb
requests
python-dotenv
pydantic